tasks: Fix [no-test]
This reverts most of #11903

When image-refresh converts an issue into a PR it renames the title and
force pushes, regardless of any of task's manipulations. When no tests
should be run upon converting, the new title should include [no-test].

Most tasks do need all tests triggered, so the [no-test] label in the
finish should be removed.

Closes #11

Closes #12

Closes #13

Closes #14

Closes #15

Closes #17

Closes #18

Closes #19

Closes #20