session: Circumvent user shell stdout noise
Commit a9375441df7 caused Cockpit logins to fail if the user's ~/.bashrc
or ~/.profile writes anything to stdout. That interferes with the
protocol and causes an error like

    incorrect protocol: received invalid length prefix

To avoid this, route the shell's stdout to stderr (so that we can still
see it in the logs) and pass session's stdout to cockpit-bridge through
fd 3. This assumes that the shell supports `>&3` input redirection,
which is the case for any POSIX shell (ash, bash, ksh, dash, zsh) and at
least tcsh.

https://bugzilla.redhat.com/show_bug.cgi?id=1710604

Closes #11978