bots: Mitigate fundamental race condition in image-refresh
This should be considered a temporary measure until the task bots
themselves call tests-scan, and not the webhook.

Remove the no-test after the task finishes again so tests can get
triggered, then trigger them manually in image-refresh after the task
has complete.

There will still be a race between the rename and the direct test
triggers, yet the probability of this going wrong should be fine until
we replace this with the task bots calling tests-scan, making the whole
thing deterministic.

Closes #59

Closes #60

Closes #61

Closes #62

Closes #63

Closes #64

Closes #65

Closes #66