bots: Mitigate fundamental race condition in image-refresh
Depending on the tests-external label inside of the image-refresh task
would depend on a race: it would expect that the task finishes and
removes no-test from the title before the webhook has called all of its
tests-scans on the direct triggers. And as the status events do not
include the pull data, this would be a race.

To avoid this race, poll for the updated head after force-pushing,
trigger all the external tests manually, and then poll that all those
statuses are present before letting the task finish and removing the
no-test from the title. This means that all external tests need to be
described in image-refresh, as it's no longer relying on the
test-external label.

Closes #71

Closes #72

Closes #73